[CGNS] FILE_PER_COMM_GROUP part 1 : Mesh OK

+++ src/Kernel/Postraitement/CGNS/Ecrire_CGNS_Links.cpp
@@ -70,14 +70,35 @@ void Ecrire_CGNS::cgns_close_solution_link_files(const double t)
 void Ecrire_CGNS::cgns_open_grid_base_link_file()
 {
   assert(Option_CGNS::USE_LINKS && !postraiter_domaine_);
-  std::string fn = baseFile_name_ + ".grid.cgns"; // file name
+  std::string fn;
 
-  unlink(fn.c_str());
+  if (Process::is_parallel() && Option_CGNS::FILE_PER_COMM_GROUP && PE_Groups::has_user_defined_group())
+    {
+      const auto& grp = PE_Groups::get_user_defined_group();
+      if (PE_Groups::enter_group(grp))
+        {
+          int nb_glob = PE_Groups::groupe_TRUST().rank();
+          envoyer_broadcast(nb_glob, 0); // XXX should do this !
+          fn = (Nom(baseFile_name_)).nom_me(nb_glob).getString() + ".grid.cgns"; // file name
 
-  if (Process::is_parallel())
-    cgns_helper_.cgns_open_file<TYPE_RUN_CGNS::PAR>(fn, fileId_);
+          unlink(fn.c_str());
+
+          cgns_helper_.cgns_open_file<TYPE_RUN_CGNS::PAR>(fn, fileId_);
+          PE_Groups::exit_group();
+        }
+      Process::barrier();
+    }
   else
-    cgns_helper_.cgns_open_file<TYPE_RUN_CGNS::SEQ>(fn, fileId_);
+    {
+      fn = baseFile_name_ + ".grid.cgns"; // file name
+
+      unlink(fn.c_str());
+
+      if (Process::is_parallel())
+        cgns_helper_.cgns_open_file<TYPE_RUN_CGNS::PAR>(fn, fileId_);
+      else
+        cgns_helper_.cgns_open_file<TYPE_RUN_CGNS::SEQ>(fn, fileId_);
+    }
 }
 
 void Ecrire_CGNS::cgns_close_grid_or_solution_link_file(const int ind, const std::string& fn, bool is_cerr)
@@ -86,7 +107,23 @@ void Ecrire_CGNS::cgns_close_grid_or_solution_link_file(const int ind, const std
   const True_int fileId = (ind == 0 ? fileId_ : fileId2_);
 
   if (Process::is_parallel())
-    cgns_helper_.cgns_close_file<TYPE_RUN_CGNS::PAR>(fn, fileId, is_cerr);
+    {
+      if ( Option_CGNS::FILE_PER_COMM_GROUP && PE_Groups::has_user_defined_group())
+        {
+          const auto& grp = PE_Groups::get_user_defined_group();
+          if (PE_Groups::enter_group(grp))
+            {
+              int nb_glob = PE_Groups::groupe_TRUST().rank();
+              envoyer_broadcast(nb_glob, 0); // TODO FIXME juste pour affichage ...a voir
+              const std::string fn_new = (Nom(baseFile_name_)).nom_me(nb_glob).getString() + ".grid.cgns"; // file name
+              cgns_helper_.cgns_close_file<TYPE_RUN_CGNS::PAR>(fn_new, fileId, is_cerr);
+              PE_Groups::exit_group();
+            }
+          Process::barrier();
+        }
+      else
+        cgns_helper_.cgns_close_file<TYPE_RUN_CGNS::PAR>(fn, fileId, is_cerr);
+    }
   else
     cgns_helper_.cgns_close_file<TYPE_RUN_CGNS::SEQ>(fn, fileId, is_cerr);
 }
